+++ Code/Effects/BuffEffects/NoclipEffect.cs
@@ -91,6 +91,21 @@ public sealed class NoclipEffect : CrowdControlEffect
 
             if (Player.HasBuff(BuffID.Shimmer) || Player.shimmering)
             {
+                // Keep shimmering if in lava
+                if (Player.IsInLiquid(LiquidID.Lava))
+                {
+                    var buffIndex = Player.FindBuffIndex(BuffID.Shimmer);
+                    if (buffIndex == -1)
+                    {
+                        Player.AddBuff(BuffID.Shimmer, 30);
+                    }
+                    else if (Player.buffTime[buffIndex] < 30)
+                    {
+                        Player.buffTime[buffIndex] = 30;
+                    }
+                }
+
+                // Prevent item use whilst shimmering
                 Player.AddBuff(BuffID.Cursed, 2);
             }
             else